+++ be/src/exec/vectorized/sorting/compare_column.cpp
@@ -293,157 +293,4 @@ void build_tie_for_column(const ColumnPtr column, Tie* tie) {
     DCHECK(st.ok());
 }
 
-struct SortedRun {
-    const Chunk* chunk;
-    std::pair<size_t, size_t> range;
-
-    SortedRun(const Chunk* ichunk, std::pair<size_t, size_t> irange) : chunk(ichunk), range(irange) {}
-
-    size_t num_columns() const { return chunk->num_columns(); }
-    size_t num_rows() const { return range.second - range.first; }
-};
-
-struct EqualRange {
-    using Range = std::pair<size_t, size_t>;
-
-    Range left_range;
-    Range right_range;
-
-    EqualRange(Range left, Range right) : left_range(left), right_range(right) {}
-};
-
-// Merge two-way sorted run into on sorted run
-// TODO: specify the ordering
-class MergeSortedTwoWay {
-public:
-    MergeSortedTwoWay(const SortedRun& left, const SortedRun& right, Permutation* perm)
-            : _left(left), _right(right), _perm(perm) {}
-
-    void run() {
-        constexpr uint32_t kLeftIndex = 0;
-        constexpr uint32_t kRightIndex = 1;
-
-        // TODO: optimize with tie
-        std::vector<EqualRange> equal_ranges;
-        _perm->resize(_left.num_rows() + _right.num_rows());
-
-        // The first column
-        equal_ranges.emplace_back(_left.range, _right.range);
-
-        // Iterate each column
-        for (int col = 0; col < _left.num_columns(); col++) {
-            const Column* left_col = _left.chunk->get_column_by_index(col).get();
-            const Column* right_col = _right.chunk->get_column_by_index(col).get();
-            std::vector<EqualRange> next_ranges;
-
-            // Iterate each equal-range
-            for (auto equal_range : equal_ranges) {
-                size_t lhs = equal_range.left_range.first;
-                size_t rhs = equal_range.right_range.first;
-                size_t lhs_end = equal_range.left_range.second;
-                size_t rhs_end = equal_range.right_range.second;
-
-                size_t output_index = lhs + rhs;
-                DCHECK_LT(lhs, lhs_end);
-                DCHECK_LT(rhs, rhs_end);
-
-                // Merge rows in the equal-range
-                while (lhs < lhs_end || rhs < rhs_end) {
-                    auto left_range = fetch_left(left_col, lhs, lhs_end);
-                    auto right_range = fetch_right(right_col, rhs, rhs_end);
-
-                    // TODO: optimize the compare
-                    int x = 0;
-                    if (lhs < lhs_end && rhs < rhs_end) {
-                        x = left_col->compare_at(left_range.first, right_range.first, *right_col, 1);
-                    } else if (lhs < lhs_end) {
-                        x = -1;
-                    } else if (rhs < rhs_end) {
-                        x = 1;
-                    }
-
-                    if (x <= 0) {
-                        fmt::print("merge left [{}, {}]\n", left_range.first, left_range.second);
-                        lhs = left_range.second;
-                        for (size_t i = left_range.first; i < left_range.second; i++) {
-                            (*_perm)[output_index++] = PermutationItem(kLeftIndex, i);
-                        }
-                    } else if (x >= 0) {
-                        fmt::print("merge right [{}, {}]\n", right_range.first, right_range.second);
-                        rhs = right_range.second;
-                        for (size_t i = right_range.first; i < right_range.second; i++) {
-                            (*_perm)[output_index++] = PermutationItem(kRightIndex, i);
-                        }
-                    }
-
-                    if (x == 0) {
-                        fmt::print("merge equal [{}, {}) + [{}, {})\n", left_range.first, left_range.second,
-                                   right_range.first, right_range.second);
-                        next_ranges.emplace_back(left_range, right_range);
-                    }
-                }
-
-                DCHECK_EQ(lhs, lhs_end);
-                DCHECK_EQ(rhs, rhs_end);
-            }
-
-            equal_ranges.swap(next_ranges);
-        }
-    }
-
-    std::pair<size_t, size_t> fetch_left(const Column* left_col, size_t lhs, size_t lhs_end) {
-        size_t first = lhs;
-        size_t last = lhs + 1;
-        while (last < lhs_end && left_col->compare_at(last - 1, last, *left_col, 1) == 0) {
-            last++;
-        }
-        return {first, last};
-    }
-
-    std::pair<size_t, size_t> fetch_right(const Column* right_col, size_t rhs, size_t rhs_end) {
-        size_t first = rhs;
-        size_t last = rhs + 1;
-        while (last < rhs_end && right_col->compare_at(last - 1, last, *right_col, 1) == 0) {
-            last++;
-        }
-        return {first, last};
-    }
-
-private:
-    const SortedRun& _left;
-    const SortedRun& _right;
-    Permutation* _perm;
-};
-
-Status merge_sorted_chunks_two_way(const Chunk* left, const Chunk* right, Permutation* output) {
-    DCHECK_EQ(left->num_columns(), right->num_columns());
-
-    SortedRun left_run(left, std::make_pair(0, left->num_rows()));
-    SortedRun right_run(right, std::make_pair(0, right->num_rows()));
-    MergeSortedTwoWay merger(left_run, right_run, output);
-    merger.run();
-    return Status::OK();
-}
-
-Status merge_sorted_chunks_heap_based(const std::vector<SortedRun>& runs, Chunk* output) {
-    return Status::NotSupported("TODO");
-}
-
-Status merge_sorted_chunks(const std::vector<ChunkPtr>& chunks, Chunk* output) {
-    if (chunks.empty()) {
-        return Status::OK();
-    }
-
-    std::vector<SortedRun> runs;
-    for (auto& chunk : chunks) {
-        runs.emplace_back(chunk.get(), std::make_pair(0, chunk->num_rows()));
-    }
-
-    int num_columns = chunks[0]->num_columns();
-    for (int i = 0; i < num_columns; i++) {
-    }
-
-    return Status::OK();
-}
-
 } // namespace starrocks::vectorized
